Put the tests for each ImageOptimizer method into its own file and move shared logic to a new Init file
