+++ Tests/Magick.NET.Tests/Shared/Optimizers/ImageOptimizerTests/TheCompressMethod.cs
@@ -1,0 +1,216 @@
+﻿using System.IO;
+using ImageMagick;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Magick.NET.Tests.Shared.Optimizers.ImageOptimizerTests
+{
+    public partial class ImageOptimizerTests
+    {
+        [TestClass]
+        public class TheCompressMethod
+        {
+            [TestClass]
+            public class WithFile : ImageOptimizerTestHelper
+            {
+                [TestMethod]
+                public void ShouldThrowExceptionWhenFileIsNull()
+                {
+                    var optimizer = new ImageOptimizer();
+                    ExceptionAssert.ThrowsArgumentNullException("file", () =>
+                    {
+                        optimizer.Compress((FileInfo)null);
+                    });
+                }
+
+                [TestMethod]
+                public void ShouldThrowExceptionWhenFileIsEmpty()
+                {
+                    var optimizer = new ImageOptimizer();
+                    using (TemporaryFile file = new TemporaryFile("empty"))
+                    {
+                        ExceptionAssert.Throws<MagickMissingDelegateErrorException>(() =>
+                        {
+                            optimizer.Compress(file);
+                        });
+                    }
+                }
+
+                [TestMethod]
+                public void ShouldMakeFileSmallerWhenFileIsCompressibleJpgFile()
+                {
+                    var optimizer = new ImageOptimizer();
+                    AssertCompress(Files.ImageMagickJPG, true, (FileInfo file) =>
+                    {
+                        return optimizer.Compress(file);
+                    });
+                }
+
+                [TestMethod]
+                public void ShouldNotMakeFileSmallerWhenFileIsUncompressibleJpgFile()
+                {
+                    var optimizer = new ImageOptimizer();
+                    AssertCompress(Files.LetterJPG, false, (FileInfo file) =>
+                    {
+                        return optimizer.Compress(file);
+                    });
+                }
+            }
+
+            [TestClass]
+            public class WithFileName : ImageOptimizerTestHelper
+            {
+                [TestMethod]
+                public void ShouldThrowExceptionWhenFileNameIsNull()
+                {
+                    var optimizer = new ImageOptimizer();
+                    ExceptionAssert.ThrowsArgumentNullException("fileName", () =>
+                    {
+                        optimizer.Compress((string)null);
+                    });
+                }
+
+                [TestMethod]
+                public void ShouldThrowExceptionWhenFileNameIsEmpty()
+                {
+                    var optimizer = new ImageOptimizer();
+                    ExceptionAssert.ThrowsArgumentException("fileName", () =>
+                    {
+                        optimizer.Compress(string.Empty);
+                    });
+                }
+
+                [TestMethod]
+                public void ShouldThrowExceptionWhenFileNameIsInvalid()
+                {
+                    var optimizer = new ImageOptimizer();
+                    ExceptionAssert.Throws<MagickBlobErrorException>(() =>
+                    {
+                        optimizer.Compress(Files.Missing);
+                    }, "error/blob.c/OpenBlob");
+                }
+
+                [TestMethod]
+                public void ShouldThrowExceptionWhenFileIsUnsupportedFormat()
+                {
+                    var optimizer = new ImageOptimizer();
+                    ExceptionAssert.Throws<MagickCorruptImageErrorException>(() =>
+                    {
+                        optimizer.Compress(Files.InvitationTif);
+                    }, "Invalid format");
+                }
+
+                [TestMethod]
+                public void ShouldMakeFileSmallerWhenFileNameIsCompressibleGifFile()
+                {
+                    var optimizer = new ImageOptimizer();
+                    AssertCompress(Files.FujiFilmFinePixS1ProGIF, true, (string file) =>
+                    {
+                        return optimizer.Compress(file);
+                    });
+                }
+
+                [TestMethod]
+                public void ShouldNotMakeFileSmallerWhenFileNameIsUncompressiblePngFile()
+                {
+                    var optimizer = new ImageOptimizer();
+                    AssertCompress(Files.MagickNETIconPNG, false, (string file) =>
+                    {
+                        return optimizer.Compress(file);
+                    });
+                }
+
+                [TestMethod]
+                public void ShouldNotThrowExceptionWhenIgnoringUnsupportedFileName()
+                {
+                    var optimizer = new ImageOptimizer { IgnoreUnsupportedFormats = true };
+                    var compressionSuccess = optimizer.Compress(Files.InvitationTif);
+                    Assert.IsFalse(compressionSuccess);
+                }
+            }
+
+            [TestClass]
+            public class WithStream : ImageOptimizerTestHelper
+            {
+                [TestMethod]
+                public void ShouldThrowExceptionWhenStreamIsNull()
+                {
+                    var optimizer = new ImageOptimizer();
+                    ExceptionAssert.ThrowsArgumentNullException("stream", () =>
+                    {
+                        optimizer.Compress((Stream)null);
+                    });
+                }
+
+                [TestMethod]
+                public void ShouldThrowExceptionWhenStreamCannotRead()
+                {
+                    var optimizer = new ImageOptimizer();
+                    using (TestStream stream = new TestStream(false, true, true))
+                    {
+                        ExceptionAssert.ThrowsArgumentException("stream", () =>
+                        {
+                            optimizer.Compress(stream);
+                        });
+                    }
+                }
+
+                [TestMethod]
+                public void ShouldThrowExceptionWhenStreamCannotWrite()
+                {
+                    var optimizer = new ImageOptimizer();
+                    using (TestStream stream = new TestStream(true, false, true))
+                    {
+                        ExceptionAssert.ThrowsArgumentException("stream", () =>
+                        {
+                            optimizer.Compress(stream);
+                        });
+                    }
+                }
+
+                [TestMethod]
+                public void ShouldThrowExceptionWhenStreamCannotSeek()
+                {
+                    var optimizer = new ImageOptimizer();
+                    using (TestStream stream = new TestStream(true, true, false))
+                    {
+                        ExceptionAssert.ThrowsArgumentException("stream", () =>
+                        {
+                            optimizer.Compress(stream);
+                        });
+                    }
+                }
+
+                [TestMethod]
+                public void ShouldMakeFileSmallerWhenStreamIsCompressiblePngFile()
+                {
+                    var optimizer = new ImageOptimizer();
+                    AssertCompress(Files.SnakewarePNG, true, (Stream file) =>
+                    {
+                        return optimizer.Compress(file);
+                    });
+                }
+
+                [TestMethod]
+                public void ShouldNotMakeFileSmallerWhenStreamIsUncompressibleGifFile()
+                {
+                    var optimizer = new ImageOptimizer();
+                    AssertCompress(Files.RoseSparkleGIF, false, (Stream file) =>
+                    {
+                        return optimizer.Compress(file);
+                    });
+                }
+
+                [TestMethod]
+                public void ShouldNotThrowExceptionWhenIgnoringUnsupportedStream()
+                {
+                    var optimizer = new ImageOptimizer { IgnoreUnsupportedFormats = true };
+                    using (FileStream fileStream = OpenFile(Files.InvitationTif))
+                    {
+                        var compressionSuccess = optimizer.Compress(fileStream);
+                        Assert.IsFalse(compressionSuccess);
+                    }
+                }
+            }
+        }
+    }
+}
